order-service/requirements.txt
kafka-python
faker # to generate fake data